requirements.txt
httpx
beautifulsoup4
lxml
pdfplumber
psycopg2-binary
python-dotenv
pandas
gspread
google-auth
ollama
google-generativeai
